ENT-6277 Add .npmrc file with enabled `engine-strict` option
Use `.npmrc` file with enabled `engine-strict` option to restrict interaction with
frontend application in wrong enviroment setup.

Add `.npmrc` file with enabled `engine-strict` to the root of the monorepo is impossible
because it throw an error for `brx-spartacus` app which has own `.npmrc` file without this option but
it goes to the root `.npmrc` during `lerna bootstrap` or `npm install`.

The `.npmrc` file with enabled `engine-strict` option has been added to the following projects:
* community/cms/editor/frontend
* community/reference-spa
* community/ui-extension
* enterprise/brie-workspace

The `.npmrc` file with enabled `engine-strict` option already exist in the following projects:
* community/brx-spartacus
* community/channel-manager/frontend-ng
* community/cms/api
* community/cms/login
* community/cms/console/frontend
* enterprise/wpm/frontend/project-management

The following projects don't have `.npmrc` file due to incompatibility with some dependencies:
* community/frontend-theme
* community/navigation-application
* community/channel-manager/ui
* community/personalization/ui
* enterprise/commerce-pickers
* enterprise/management-apis/ui